Let Battenburg constraints takes care of cells being different.
But only sometimes. We already have automatic differences when
the parity differs, but this is different on the diagonal. If
the Battenburg square lies all in the same box, the values on
the diagonal must differ. If the Battenburg square straddles two
(or four) boxes, the values on the diagonal may be the same
(unless there is an anti-king or diagonal constraint).