Utils: Replace cell_name() by cell().
The latter just checks %args, which callers pass on from their args,
then just do "the right thing", to generate the cell name.
Hence, we no longer require other subs to extract the row and column.

Used parameters:

    - cell           (name of cell)
    - row            (row number of cell)
    - col or column  (column number of cell)

All other parameters will be ignored. The returned cell name will
always be of the form "RddCdd". If we cannot create an acceptable
cell name, we will thrown an exception.

For backwards compatability, for now, we keep a cell_name() function
as a wrapper around cell().